tools/c7n_org - fix worker count type when set via env var (#2113)

+++ tools/c7n_org/c7n_org/cli.py
@@ -47,7 +47,7 @@ from c7n.utils import UnicodeWriter
 log = logging.getLogger('c7n_org')
 
 
-WORKER_COUNT = os.environ.get('C7N_ORG_PARALLEL', multiprocessing.cpu_count() * 4)
+WORKER_COUNT = int(os.environ.get('C7N_ORG_PARALLEL', multiprocessing.cpu_count() * 4))
 
 
 CONFIG_SCHEMA = {
